Bugfix: The primary account should not be guessed by mfcs_users_load_banner_id_to_user_id()
The synchronize user function will call mfcs_users_load_banner_id_to_user_id() and use its results to determine if the primary account needs to be auto-assigned.
If mfcs_users_load_banner_id_to_user_id() defines a primary account when the the data is NULL (instead of 0 or 1), then the caller incorrectly thinks that there is no need to assign a primary account.
When the primary account setting is NULL, it must be set to 0 or 1.