Feature: provide addition regarding technical requests
As per request, add additional text for when technical equipment is requested.

The code contains a commented out version containing a link to a page on www.mcneese.edu.
That code will not be used until such time www-1360 is approved and an appropriate page is created to link to.

ticket: fcs-75
ticket: www-1360