Update: METHOD:PUBLISH should appear at message level and not event level and content-type header must have additional data
According to the standard, when method:publish is provided, it must also have a content-type setting including that method.
In addition to specifying the method in the content type, change the type from text/plain to text/calendar.

It seems that the component="vevent" may also be required.

see: https://tools.ietf.org/html/rfc5545#section-3.7.2