Bugfix: simultaneous requests, with 1 later locked, needs to lock out all others (part 1)
This is the first part of the solution to preventing a unlocked requests from becoming locked when there already is a locked requests.

Example:
  Request A is created, when there is no locked request (allowed)
  Request B is created, when there is no locked request (allowed)
  Request C is created, when there is no locked request (allowed)
  Request D is created, when there is a locked request (denied and auto-closed)

  Requests A through C are the problem.
  If Request A is set to the Locked state, then B and C should no longer be allowed to be locked.
  The problem is that the code was still allowing both B and C to be locked, resulting in 3 separate reservations for the same room.

This is the first part of the solution that simply blocks further progress of an unlocked request if the room is no longer available.

ticket: fcs-77