Cleanup: form_set_error() requires a missing extra ']'
The actual form id in this case is exactly 'review[operations][decision]', but the way drupal handles form set error requires the format to instead be: 'review][operations][decision'.
This is confusing, but it is likely because they use the explode() php function on the string ']['.
Drupal will later convert that to the expected 'review[operations][decision]' format automatically.