Update: make the holidays conflict informative instead of restrictive
Its probably a better idea to just inform users and let the Venue Coordinators and all the rest handle whether or not something should happen on a holiday.

This adds additional warnings for reviewers during appropriate review steps in the Review Operations area to improve odds that a reviewer is aware of the holiday.

ticket: fcs-12