Bugfix: rewrite mfcs_auto_block_accounts_cron(), increase performance, improve logic, block correct accounts
The function was incorrectly blocking almost all accounts.
The function was very slow.
There are additional exception cases that were also not being processed and should be.

This now avoids expensive calls and as a result except the cache to refreshed prior to the usage of this function.

This should be followed up with additional login tests.
On login, especially if the account is locked, check to see if the account has banner roles and unlock it unless it is locked with 'locked exception'.
- Note: 'locked exception' role will be created following this commit.