Bugfix: use HTTP Post instead of HTTP Get for datatables ajax requests
The default behavior of datatables is to use HTTP Get requests.
This results in a large amount of data being effictively passed via the url in the request.
This then floods the logs with extra data, wasting disk space.

The HTTP Post methid is generally the proper and more appropriate method to use.
Change the behavior to use HTTP Post as per datatables settings.