Bugfix: don't add accordion styles from msugrid subtheme to the main theme
These styles are applied globally instead of being based on a custom class name.
This results in affected other parts of any site that uses this theme in a negative way.

Move the custom accordion changes back to the sub-theme.
This code will still cause problems but at least those problems will now be limited to the sub-theme.