Cleanup: remove manually added \n on HTML markup
I consider it good practice to use \n to ensure clean and consistent code.
Browsers tend to interpret the \n in certain conditions (such as with display: inline-block) and cause problems with the display.

To avoid this, I am forced to abandon the idea of adding \n at the end of markup.