Bugfix: make sure fees remain when set to university business
The behavior did not change with university business clearing use fees.
What changes was when the use fees were moved to the requirements.
The University Business was set before use fees were assigned and so this "problem" was never noticed.

Only the Facilities Fees were waived when set to University Business.
Change the behavior to simply not require Facilities Fees.
Change the behavior to not clear or unset the use fees.

A bug was noticed where $requirements_university_save was being used where $requirements_waived_save is the obvious correct choice.

Added comment about some of the "Waived" code because that code may or may not be used anymore (it may be stale code).

ticket: fcs-79