Bugfix: explicitly grant administer access
I noticed this when I change the admin drupal role to only enable the admin options.
I then set my admin account to only have the admin role and not the manager role.
This revealed to me that the admin permissions were not explicitly specified under a number of circumstances.

Make sure the administer option is checked last as a trivial performance optimization.

There are some permissions that are expected to be manager only and not directly control by admin.
- Specifically, the review operations.