Bugfix: reviewer cannot see log when they are the requester but not allowed to review
Example:
- user_1 has role "reviewer".
- user_1 is not assigned access to review request_1.
- user_1 is the requester or request_1 (created the request).

In the code, the reviewer access is processed first and also processed exclusively.
Because of  that, user_1 gets an access denied because they cannot review the request despite being the requester.

This changes the behavior in the following ways:
1) For the review 'log', process the requester roles before processing the reviewer roles.
2) If access is not granted, then process the reviewer roles.

This should have a conditional performance gain because the reviewer role access checks are expensive and the requester role access checks are inexpensive.