Update: grant managers more complete access to the review requests listing
The previous behavior, review requests was displayed based on a persons current priveleges as a reviewer.
The system managers did not get treated any differently.

To get this listing the manager would have to set themselves up as a proxy.

Instead of doing all of that, make the managers life easier by granting them full access to all requests that can possibly appear under the Review Management listing.
This is a change in behavior, so this will be followed up with the ability to goggle "view as Manager" and "View as Reviewer" button in the same way "Show My Requests" is implemented on the Request listing page.

ticket: fcs-79