Feature: grant managers the ability to add/remove users (and related bugfixes)
Added a new interface on the management page called "Manage Users".
- Provides a simplified user management interface.
- Grants managers the ability to add and update users on the system (but they cannot do anything to administrator accounts).

This forces considerable restructering of some of the code for improved organization.

The user login and role generation processed was updated and fixed.
- There were inconsistencies in how roles were being assigned and handled.
- The appropriate $edit['role'] data now treats the array values as the role ids instead of role names to be consistent with drupal core role updates.

During the insert process, processed information would get overwritten.
- This was caused by other categories calling the user hooks (specifically ldap was causing this).
- Restricted the user hooks to being called only when the category is 'account'.

Note: this is incomplete and I should not push this commit until finished.
- If you see this, then I accidentally pushed this commit!