Feature: add request managers interface and implement override support
The number of options being added to the view page is getting larger, so just add a manage button (and therefore interface).
Implement the override support inside of this manage interface.
Move the Re-Assign functionality into the new manage interface.

Note: this manage interface is a per-request manage interface (not to be confused with the website-level manage interface)!

ticket: fcs-79