Bugfix: conditionally allow for javascript attributes (add support for text_format filtering)
The security settings were removing js attributes for the ns form.
A new filter (called full_html_js) was created to address this.
When that filter is in use, do not auto-remove the javascript attributes.

This format should be used with care because it can be abused and can become a security problem.
This format should only be used in exceedingly rare circumstances.

ticket: www-1474