Update: also check the room availability during the individual date validation process
The behavior of the code had to change where the specific function being updated used FALSE to represent individual times for all day events.
The behavior of the rest of the system is using NULL.
Change the behavior of using FALSE to the behavior of using NULL.

I did not check to see if the FALSE behavior is used anywhere else.