Update: rework the review process to allow re-submitting the review operation
The reviewer can now re-submit their review and change their position.
For examole, during the review process when the reviewer reviewed as "No Issue", something may have come up and they now need to say 'Issue'.

The review permissions and access are tied to this process and are now removed.
- This concerns me in that I may have overlooked something (keep an eye out for regressions).

ticket: fcs-79