Bugfix: uncancelled requests get auto-cancelled again
Whenever the auto-cancel cron job is run, it will auto-cancel all requests.
In some cases, a request needed to be uncancelled but before anything could be done the auto-cancel cancelled it.

This adds an interface for locking particular aspects of a request.
An auto-cancel lock is now added such that when the lock is enabled/true then auto-cancel will be prohibited.

When a request is manually uncancelled, the auto-cancel lock will automatically be set to true.
The end result is that manually uncancelled requests should never become auto-cancelled again.

Additional controls are also provided in the new request management interface.