Regression: the continue button on the request form was not processing all of the validation conditions
The previous implementation fixed the immediate problem but prevent the form from continuin.
Re-structure the create process to ensure that the proper validation occurs and the form can still be continued.

This adds an alert and a watchdog error message for any unhandled cases that I may have overlooked with these changes in hopes of catching any further regressions.