Bugfix: errors detected when creating quick meetings
In particular, I noticed that the decision options of "Issues" or "No Issues" appears for completed requests.
I have a vague feeling that this was intended, but because I do not remember and so I am disabling this.
Disabling "Issues" and "No Issues" for completed requests makes sense to me.