Bugfix: all day request problem detection query not including non-all day unavailabilities
When a request date is set to all day, the unavailability detection is failing because the test only checked for the midnight today test as opposed to the range.
This happened because there are no 'tomorrow' fields for all day.
The solution is to add 86400 to the all day column to catch all times within that date.