Feature: add logical data defragmentation support, starting with user ids
Provide the ability to logically "defragment" the user ids.
This is dependent on the algorithm and in this case there are two algorithms.
Both of these are meant to randomize the user ids for security improvements.
The second algorithm will move all locked/deleted accounts to the higher numbers.
In both bases, the first 100 user ids are reserved for special system accounts.