Bugfix: request fields should be sorted by delta on load
When a date category was replaced in the database, the category swapped with the category of another delta.
Example:
  Correct Order:
    [0] (Event/Meeting) 2016/01/01
    [1] (Rehearsal/Setup) 2016/01/02
    [2] (Cleanup/Breakdown) 2016/01/03
  Observed Order:
    [0] (Rehearsal/Setup) 2016/01/01
    [1] (Event/Meeting) 2016/01/02
    [2] (Cleanup/Breakdown) 2016/01/03

This is happening because the deltas are not being sorted and the database is loading the rows in whichever order it feels like.
The date day field and the date category field are two separate tables that should be merged with their revision and delta.