Bugfix: primary_account and ldap_duplicate postgresql boolean types may need to be converted to the words TRUE/FALSE
I noticed that when passing the boolean TRUE, it gets properly converted into a format that postgresql can handle.
However, when passing the boolean FALSE, only an empty string is supplied (causing an SQL error).

To prevent this problem, just convert both boolean values to the appropriate TRUE/FALSE string.