Bugfix: other roles disappear when searching users by a role name
Performing a where condition on an array_agg() causes all other values to disappear no matching that condition to not be joined.
This means that if I search for 'reviewer', only the reviewer role will appear.
This is confusing as the user page is intended to show all roles, not just the searched for row.

To solve this, perform the search using a subquery.