Cleanup: remove 'create new user' functionality from 'venue coordinator proxy' page
This was added early on before I had a proper user add functionality implemented.
This add functionality is essentially a hack and is now no longer needed that a more proper way to add users is implemented.

Removing this also helps avoid confusion by placing the add user functionality in fewer places.