Bugfix: improve mfcs_synchronize_user_settings() and related based on testing
The problems code was originaly using the now 'updated' column for reporting when problems were updated and not when any changers to the user were made.
When I re-designed the 'mfcs_users' table I changed the behavior of that column.
This adds a new column called 'updated_problems' to preserve the previous behavior.

The mfcs_synchronize_user_settings() function is not performing the invalid banner check when $refresh_banner is set to FALSE.
This check does not require doing a refresh, so perform the check manually.

The mfcs_synchronize_user_settings() function calls itself recursively, but it is loading a user object from a user id and passing the user object.
This can be made more efficient now that a user id may be directly passed instead of a user object.