Update: add support for checking user problems for individual users through interface and fix any observed problems
Doing the individual checks allowed me to easily check to the code for problems.
This fixes the logic problems and other mistakes that I noticed.

In particular, I had to re-write how the user_save() user argument was being defined to ensure that the roles get properly updated.
The previous behavior did not consistently work as expected.

I also noticed that the ldap_duplicate and primary_account fields were not being processed correctly and were not even being updated.
This change required providing more information from the caller.

Now, if multiple primary accounts are only counted as a problem when the accounts are enabled.
Now, if there is no primary account, even if there is a disabled account, a default is provided (by the information function, this is not necesserily assigned).