Bugfix: holidays warnings (and possibly unavailabilities) are not always displayed and show warnings/unavailabilities on complete request views
The $all_holidays variable was being overwritten instead of being appended to.
If a date following a holiday date was called, it would set the $all_holidays to an empty array, thereby hiding the hiding and preventing any kind of warning.

Make sure the warning is always displayed for completed and approved requests.