Bugfix: Re-Process Reviewer/Venue-Coordinator Role option is incorrectly removing the problem id
When the user had a reviewer role assigned, the reviewer not allowed problem flag gets set.
This problem flag should only be removed if it is reset or the user no longer has the reviewer role specified.
That said, when this flag is first assigned, the reviewer role is already removed.
Therefore subsequent tests would auto-remove this and hide the fact that the reviewer role not allowed problem was ever encountered.

New behavior:
- If the reviewer role gets unassigned by the check or is now allowed durong the check, then and only then remove the reviewer not allowed problem flag.
- The reviewer not allowed problem flag may still be reset like any other flag.

I also re-worded of the reviewer not allowed flag text.