Bugfix: variables having random (invalid) data on review log view
This turns out to be a bug resulting from PHP variable names in the for loop.
Because the loops generated values were used with references (ampersand) and the same variable name was used later, unrelated operations outside the loop ended up modifying the contents of the array in the loop.

This is solved by unsetting all array variables to ensure this kind of accident does not happen.