Bugfix: make sure copy operations do not process the edit operation restrictions
When a Quick Meeting request is copied by a requester who is not allowed to create a Quick Meeting, the Quick Meeting option should not appear.
Repeat the same tests for the request classification.

During the request classification option, if the user who copies has only one classification, auto-provide that as the default classification.