Bugfix: users with 'bannerless exception' role are not being properly tested
When the 'bannerless exception' role is assigned to a user, the mfcs_synchronize_user_settings() function will immediately exit.
Instead, when 'bannerless exception' role is assigned, just avoid synchronization.

This allows for performing all of the existing tests in case the user with the 'bannerless exception' role is given a banner id.
- This often happens when there is a bug or issue with an account and 'bannerless exception' is assigned as a temporary workaround.