Bugfix: user_save() is not always correctly settings the banner/bannerless information and avoid calling user_save() on synchronize
There were a few problems with bannerless exception and other unique situations where the roles were being improperly saved or not saved at all.
I noticed that the array_diff() function does not work very well and I abanonded its use.

The reviewer role was not always being added when the user is a proxy or the user is explicitly a reviewer.

The synchronization may need to be called multipled times because the information may have changed, so remove the drupal static cache.

The synchronize function needs to be called after user_save().
If synchronize calls user_save() and changes are made, then synchronize would need to be called again.
Avoid this by directly updating the user roles instead of calling user_save() from within the synchronize function.