Bugfix: some duplicate accounts have problems with roles being assigned
This happens because some duplicate accounts are not properly associated with the banner id.
The banner id is currently being associated with a single user id.

This changes the behavior to have a banner id to banner roles cache table.
This table can then be used to determine a users banner roles independent of a specific user.

This requires the following SQL query:
  start transaction;

  create table mfcs_banner_event_roles_to_banner_ids (
    banner_id bigint not null,
    event_role_id bigint not null,
    date bigint not null,
    disabled smallint not null default 0,

    primary key (banner_id, event_role_id),
    constraint mfcs_banner_event_roles_to_banner_ids_banner_id_check check (banner_id >= 0),
    constraint mfcs_banner_event_roles_to_banner_ids_event_role_id_check check (event_role_id >= 0),
    constraint mfcs_banner_event_roles_to_banner_ids_date_check check (date >= 0),
    constraint event_role_id_fkey foreign key (event_role_id) references mfcs_banner_event_roles(id) on update cascade on delete cascade
  );

  grant select, insert, update on mfcs_banner_event_roles_to_banner_ids to fcs_user;

  commit transaction;