Workaround: temporarily disable amendment operations
Users were avoiding making changes to the request because the amend process was causing a reset that they wanted to avoid.
Disable this functionality to avoid doing further work as a simple work around for this issue.
This may be re-enabled in the future based on observations.

Ideally, this should be re-written to only trigger amendments (and therefore review restarts) when only the following fields change:
- rooms/venue change
- dates