Cleanup: reword stale requests reports to be about one or more dates instead of all dates
The original statement was true at one point in time but due to all of the changes it is not necessarily true.
Stale requests are triggered if one or more dates are in the past but the request is not cancelled.
This is not necessarily a problem because requests span multiple days may have some parts while others are upcoming.

This re-words the statements.