Bugfix: managers should not be able to specify issues/no-issues
Manager is not a classification for the purposes of the review issues/no-issues process.
Strip out the relevant code and make sure managers cannot specify issues/no-issues.

This effectively removes a feature where a reviewer only reviews as classifications they have not already reviewed as.
That feature also caused a problem when reviewers reported an issues/no-issues during the final decision step but because they had already reviewer it was being set to manager!
(Because I had the manager user in use at the time.)