Feature: add support for club meeting type as a work around to relax the 3-week requirement for student organizations
Provide a new meeting sub-type that is available only to student organizations.
When this meeting type is selected, the 3-week requirement should not be enforced.
Only a warning will be presented.

This is currently being discussed following the meeting about the changes from fcs-79.
This may end up getting disabled in the future.

ticket: fcs-79