Bugfix: requester, manager, and administrator classifications are not appearing properly in the review log
They are not being saved to the database and a dynamic detection is being used.
Because of dynamic detection a person could have made a comment as a non-admin and then later made an admin.
The comment from that point forward would reference the user as an admin (which was incorrect at the time of the comment).

Both manager and administrator are only added if the user does not belong to any other classification at the time of the review operation.