Bugfix: handle cases where requester and venue coordinator should not appear in the review log
Users with multiple relevant classifications, such a venue coordinator who makes a reqest making that user both a venue coordinator and a requester, can perform more tasks than some of the roles require.
In the case of a requester, they should only be able to make comments so do not log any other action as a requester.

This adds 'never_requester' and 'never_coordinator' options to the mfcs_workflow_set_requester() function.

I also added an @fixme comment for an area where mfcs_workflow_step_review() probably shouldn't be called.