Bugfix: further review log and reviewer classification changes, focusing mostly on financial and insurance (additional issues, see notes)
The financer and insurer roles should be reported as relevant in the workflow log.

Fix some consistency issues with manager and administrator roles in workflow log.

Add a new "financial" and "insurance" classification, requiring the following SQL query:
  start transaction;
  insert into mfcs_taxonomy_reviewer_classification  (id, machine_name, human_name, hidden) values (21, 'financial', 'Financial', 1);
  insert into mfcs_taxonomy_reviewer_classification  (id, machine_name, human_name, hidden) values (22, 'insurance', 'Insurance', 1);
  alter sequence mfcs_taxonomy_reviewer_classification_id_seq restart 23;
  commit transaction;

I am currently waiting on feedback on how to handle the insurance and requirements "rate" situation.
I added temporary code and appropriate comments that will need to be fixed, updated, or otherwise altered once a decision is made.

In particular, rate is currently tied to the other requirements and when other requirements are updated, so is rate (even if there is no change from the default).