Workaround: use absolute url paths instead of local paths for PDF generation
For some reason relative paths are not working.
This is most obvious when using images (because they are not displayed in the PDF).
By switching to absolute paths (and making sure /etc/hosts points to the ip address instead of 127.0.0.1 for the appropriate host), the images now appear.

It is entirely possible that this is one of the reasons why I had to manually embedd css and js.
The files may simply not have been added properly with relative paths.
This is something to look into because if the css and js does not have to be generated, then the content should need fewer alterations.