Update: improve performance using defer loading, also removing no_js_fallback
When a page load the list gets loaded and then an ajax call loads the list agaon.
To avoid this, I did not load the list, just the list count and then called the ajax load.
This still required an extra transaction that took maybe an extra 300ms to load the table.

With defer loading, I can specificy the list, pre-load the data and not need ajax to load after the initial page load.

Ajax is still used after the initial page load for all dynamic transactions such as searching and changing page numbers.

With this change, the no_js_fallback functionality is no longer needed.