Bugfix: no-script and script problems involving datatables and displaying of column headers and rows
When javascript is disabled the columns are not always displayed properly.
Sometimes there is also a flicker.

This is happening because the appropriate class names are not appearing when javascript is not available.
(which also happens on page load before javascript is loaded and alters the original data).

This introduces a trivial problem where some class names are duplicated (which cannot be avoided currently due to datatable's design).

I also noticed some cases where script-hidden is being used and should not be.