Bugfix: experimentally disabled custom script-hidden css class adding code
The datatables is having certain invisible columns have 'script-hidden' added to them.
When showing/expanding a column, the column header is not appearing.
Removing this 'script-hidden' solves the problem.

The problem here is that I do not remember why I added it.
I need to research the git change logs and try to find out.
Until then, I believe it makes more sense to disable the problematic block of code.