Update: implement the fullcalendar monthly calendar and fix misc problems
This adds support for using the custom fullcalendar instead of the original calendar.
To prevent confusion and avoid having to re-teach users, this is added as a customization option that users must manually set.
This is considered experimental but the goal is to use FCS as a testing ground for further work on production (specifically in regards to www-171).

One of the problems that was fixed was that the original calendar was not displaying holidays on the dashboard.

ticket: www-171