Update: further tweak fullcalendar and fix related problems
Remove extra initialization on javascript items that resulted in preventing the holidays from appearing.
Prepend 'holiday-' to the holiday id to ensure that it does not conflict with event ids.
Make sure current day is not highlighted on day view, because that makes everything yellow and its annoying.
Clicking day links on the month calendar now goes to the per-day calendar.
Add missing location, building, and room class id tags.