Update: unavailability pages should accept the unavailability id if the rest is not known
This fixes a problem where when specify a year, such as (/requests/unavailable-0/view/2016) an access denied page for a different function is called.
This is fixed by calling the custom unavailability redirect function when there is only 1 url argument for view, edit, and delete operations.

This allows for an unavailabilty id to be supplied in the url without knowing the year or the room.
A redirect will be performed only if the unavailability exists, is not deleted, and the user has the appropriate access priveleges for that unavailability.