Bugfix: multiple times in full calendar are not consistently appearing on each request instance
The tooltip contents of each request with multiple times is not consistently showing all of the times.
This happens because not all the times are known while processing the loop.

Adding a pre-process allows for determining all the times and storing these times in a single string to be used on each time instance during the main loop.