Update: add functions for processing database timestamps
All this does is add the functions.
These functions are not yet used.

The reason behind this is that PHP stores unix timestamps in UTC but calculates and generates them using localtime!
This causes two different timezones for the same UTC time to be generated as two different UTC timestmaps.
They should be identical!

For example, 'Midnight, January 1st, 2018 CST' should have the UTC string of: 1514764800.
Instead a value of 1514786400 is generated, which is the UTC plus an additional 6 hour offset.
This means the UTC is a hybrid UTC/Localtime and fails to actually represent UTC!