Bugfix: user roles are being overwritten during some function calls that do not and should not alter roles
Drupal inconsistently provides the value of the user roles to be either a string describing the role or a number representing whether or not that role is enabled.
When a string, disabled roles do not exist.
When a number, disabled roles or roles set to 0 do not exist.

Handle the case when the ldap name is invalid or ldap otherwise returns false instead if incorrectly initializing user settings.
- Note: this may reveal a bug in new accounts.
        check to see if new accounts are properly initialized.

Remove some code involve roles that was not being used (aka: dead code).