Feature: add support in fullcalendar to apply current day css affects
Fullcalendar does not provide a css class to assign the current day to all fields.
It only does this for a small subset.
This only gives partial control over styling the css.

Due to the layout/structure of the generated fullcalendar markup, the css attribute :last-of-type cannot be used.
Instead, a new class called fc-event-container-last, has been added.