Bugfix: use block name for login black instead of block id
Using block id is different and also appears to be assigned on a per row basis in the sort order.
This may be a bug in how the block id css is generated (needs investigation).
Using the block name is more distinct and consistent between systems for the user login block.