Update: improve default calendar processing and handling
This also makes the default full calendar.
This provides a way to make site specific default calendars (via settings.php).

This makes the default for fcs production to be the original calendar.
This is added as a failsafe in case I forget to set all existing accounts to explicitly use original calendar instead of default.
If I do remember, then I should need to then remove the default calendar override from original to full so that all new accounts use full calendar as the default.