Feature: add a more granular banner id and banner role problem detection, part 2
This part focuses on support for LDAP issues.

Users cannot normally login if they are not in LDAP.
However, settings may change and some accounts may be manually added.
These cases may require detection of LDAP issues.

This is a check to see if the banner ID is available in LDAP for the specified user.

ticket: fcs-97