Update: utilize the newer search by date where possible and fix any related issues
The structure of the requests listing and review listing must join the field dates and times, which takes too long.
The best solution for that is to build a custom table that is auto-populated via some sql trigger on update.
This custom table can then be joined as the dates fields table, with the dates created as an sql array.