Bugfix: (Corner Case) Some Venue Coordinators are not assigned the Reviewer Role when they otherwise should be
It seems that only the Room Coordinator list was being processed when both the Venue and the Room Coordinator lists should be used for detecting whether or not a user should have the Reviewer Role.

This adds two new functions and corrects a strange problem in the existing function.
I suspect that this function was not fully completely and I forgot all about it.

ticket: fcs-98