Bugfix: comments done by reviewers outside of their normal review step does not show their classification
The failsafe "none" ends up getting used, and a warning like the following is sent to the watchdog:
  "Failed to assign any classifications for the review step 3 with decision 11 for request 1234 for the user 123"

To prevent this, if the user has no valid classifications for the current step, attempt to use any existing classifications from either the review or the final decision steps.
This will preserve classifications for comments across steps.

ticket: fcs-98