Bugfix: verbage on auto-cancel 'locking' functionality improvements and auto-cancel save improvements
The auto-cancel verbage is sufficiently confusing that I am considering this a bug.
The verbage was change to make more sense and hopefully be less confusing.

I happened to notice that when set to 'always' (which used to be 'default' prior to this commit), the corresponding table entry was not properly deleted.
The default behavior should not populate the mfcs_request_locks table.