Bugfix: 'first time' should be default instead of 'always' for auto-cancel lock
When I implemented 70cd097c9609b8a0ab53e3042401694876cfa8a4, I overlooked that the default behavior is to automatically set 'prohibited' after a request is uncancelled.
This essentially creates the 'first time' situation.

The default behavior is therefore 'first time' and not 'always'.
This changes the verbage to reflect that functionality.

Signed-off-by: Kevin Day <kday@mcneese.edu>